backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
flask-jwt-extended==4.5.3
flask-sqlalchemy==3.1.1
flask-migrate==4.0.5
anthropic==0.37.1
requests==2.32.3
python-dotenv==1.0.0
werkzeug==3.0.1
reportlab==4.0.7
bcrypt==4.1.1
python-dateutil==2.8.2
langchain-core==0.3.15
langchain-community==0.3.3
langchain-openai==0.2.6
langchain-anthropic==0.2.4
langgraph==0.2.40
langsmith==0.1.129
fastapi==0.115.0
uvicorn==0.32.0
pydantic==2.9.2
chromadb==0.4.22
langchain-text-splitters==0.3.2
pypdf==4.0.1
python-docx==1.1.0
markdown==3.6
tiktoken==0.8.0

